Use cred file reader from SDK. Remove write for cred file interface.
