+++ package.json
@@ -245,7 +245,7 @@
     },
     "dependencies": {
         "async-lock": "^1.1.3",
-        "aws-sdk": "^2.317.0",
+        "aws-sdk": "^2.395.0",
         "cloudformation-schema-js-yaml": "^1.0.1",
         "del": "^3.0.0",
         "fs-extra": "^6.0.1",
